Create second_input
Given a string. Find the second occurrence of the letter f in this line and print the index of this occurrence. If the letter f appears in this line only once, print the number -1, and if it does not occur even once, print the number -2. When solving this problem, you cannot use the count method.