Create bombs
The first line introduces the number n - the number of villages (1 <= n <= 100000). The second line contains n different integers, the i-th of these numbers sets the distance from the beginning of the road to the i-th village. The third line of the input data sets the number m - the number of bomb shelters (1 <= m <= 100000). The fourth line contains m different integers, the i-th of these numbers sets the distance from the beginning of the road to the i-th bomb shelter. All distances are positive and do not exceed 10⁹. Settlement and shelter can be located at one point. Print n numbers - for each village print the number of the nearest bomb shelter. The bomb shelters are numbered from 1 to m in the order in which they are specified in the input.