Create polyglots
Each of the N students of a school knows Mᵢ languages. I know that everyone knows schoolchildren.

Input format

The first line of input contains the number of students. Then there are N numbers Mᵢ, after each of the numbers there are Mᵢ lines, the names of languages ​​that the 1st student knows. The length of the names of languages ​​does not exceed 1000 characters, the number of languages ​​is not more than 1000. 1≤N≤1000, 1≤Mᵢ≤500.

Output format

You know the number of languages ​​that all students know. In the second row is a list of such languages. On the following lines is a list of such languages.